refactor: split oversized modules into focused sibling files
